requirements.txt
APScheduler==3.8.1
# apturl==0.5.2
# asgiref==3.4.1
# asn1crypto==0.24.0
# backports.zoneinfo==0.2.1
boto3==1.19.7
# botocore==1.22.7
# Brlapi==0.6.6
# certifi==2018.1.18
# cffi==1.15.0
# chardet==3.0.4
# click==8.0.3
# colorama==0.3.7
# command-not-found==0.3
# cryptography==36.0.0
# cupshelpers==1.0
# dataclasses==0.8
# defer==1.0.6
# Deprecated==1.2.13
# distro-info===0.18ubuntu0.18.04.1
Django==3.2.8
django-apscheduler==0.6.0
django-ckeditor==6.1.0
django-cors-headers==3.10.0
django-js-asset==1.2.2
django-multiselectfield==0.1.12
django-storages==1.12.3
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
firebase==3.0.1
# Flask==2.0.2
# gcloud==0.17.0
# googleapis-common-protos==1.53.0
# httplib2==0.9.2
# idna==2.6
# importlib-metadata==4.8.1
# importlib-resources==5.4.0
# itsdangerous==2.0.1
# Jinja2==3.0.2
# jmespath==0.10.0
# jsonfield==3.1.0
# jwcrypto==1.0
# jws==0.1.3
# keyring==10.6.0
# keyrings.alt==3.0
# language-selector==0.1
# launchpadlib==1.10.6
# lazr.restfulclient==0.13.5
# lazr.uri==1.0.3
# louis==3.5.0
# macaroonbakery==1.1.3
# Mako==1.0.7
# MarkupSafe==2.0.1
mysqlclient==2.1.0
# netifaces==0.10.4
numpy==1.19.5
# oauth==1.0.1
# oauth2client==3.0.0
# olefile==0.45.1
pandas==1.1.5
# pexpect==4.2.1
#Pillow==5.1.0
Pillow==6.2.2
# protobuf==3.19.1
# pyasn1==0.4.8
# pyasn1-modules==0.2.8
# pycairo==1.16.2
# pycparser==2.21
# pycrypto==2.6.1
# pycryptodome==3.4.3
# pycups==1.9.73
# PyGObject==3.26.1
PyJWT==1.7.1
# pymacaroons==0.13.0
# PyMySQL==1.0.2
# PyNaCl==1.1.2
# Pyrebase==3.0.27
# pyRFC3339==1.0
# python-apt==1.6.5+ubuntu0.7
# python-dateutil==2.8.2
# python-debian==0.1.32
# python-jwt==2.0.1
# python-magic==0.4.16
# pytz==2018.3
# pytz-deprecation-shim==0.1.0.post0
# pyxdg==0.25
# PyYAML==3.12
# razorpay==1.3.0
# reportlab==3.4.0
requests==2.11.1
# requests-toolbelt==0.7.0
# requests-unixsocket==0.1.5
# rsa==4.7.2
# s3cmd==2.0.1
# s3transfer==0.5.0
# SecretStorage==2.3.1
# simplejson==3.13.2
# six==1.11.0
# sqlparse==0.2.4
# sseclient==0.0.27
# system-service==0.3
# systemd-python==234
# twilio==7.4.0
# typing-extensions==3.10.0.2
# tzdata==2021.5
# tzlocal==4.1
# ubuntu-advantage-tools==27.8
# ubuntu-drivers-common==0.0.0
# ufw==0.36
# unattended-upgrades==0.1
# urllib3==1.26.7
# usb-creator==0.3.3
# wadllib==1.3.2
# Werkzeug==2.0.2
# whitenoise==5.3.0
# wrapt==1.13.3
# xkit==0.0.0
# zipp==3.6.0
# zope.interface==4.3.2
